polmatrix-etl/requirements.txt
pandas==2.1.4
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
wbdata==0.3.0
sqlalchemy==2.0.23
schedule==1.2.1
python-dateutil==2.8.2
numpy==1.26.2
lxml==4.9.3
beautifulsoup4==4.12.2
jsonref==1.1.0